refactor events with reducer
