+++ src/components/App.js
@@ -9,7 +9,10 @@ import reducer from '../reducers'
 console.log({AppContext})
 
 const App = () => {
-  const [state, dispatch] = useReducer(reducer,[])
+  const initialState = {
+    events: []
+  }
+  const [state, dispatch] = useReducer(reducer, initialState)
 
   return (
     <AppContext.Provider value={{state, dispatch}}>
